requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
diskcache==5.6.3
distro==1.9.0
Django==5.1.5
django-allauth==65.3.1
django-allauth-ui==1.5.3
django-widget-tweaks==1.5.0
Flask==2.3.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
langcodes==3.5.0
language_data==1.3.0
llama_cpp_python==0.3.16
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.1
oauthlib==3.3.1
openai==1.96.1
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.32.0
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
slippers==0.6.2
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typeguard==2.13.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2